requirements.txt
pymongo==3.9.0
python-dateutil==2.8.1
tweepy==3.8.0
